requirements.txt
requests
setuptools

python-dotenv
pytest